code_samples/multi_agent/03_performance_optimization/requirements.txt
# Core dependencies
python-dotenv==1.0.0

# LangChain and AI dependencies (optional - will fall back to mock agents if not available)
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.12.0

# Web search tools (optional)
duckduckgo-search==4.4.0

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Monitoring and logging
structlog==23.2.0